introduced swap calculator to avoid namespace HiveMarket usage
